[CLN] Refactor test imports and file structure
Fixed import paths in `test_plot_3d.py` for `check_image_hash`. Moved `util_tests.py` under the appropriate `tests` directory for better organization.
